Increase Erase Block Timeout
Increase the erase block timeout to 3000 milliseconds to allow slow
PCMs time to erase a section of flash and return properly. This change
requires using a newer STN specific command to set the timeout that isn't
limited to 1020 milliseconds.

The App currently waits 3 seconds anyway for a response, the timeout
from the programming device should match it to have the best chance of
returning success if not at least a status indicator as to why it failed.